ASU-1002: Move services to a package
Split the services module into a package, so that the module won't
become massive when HITAS and HASO application services are added.
